axes_calibrate: Wait for movement to stop rather than just sleep
We cannot just sleep()/delay(), ahrs_update needs to be called
continuously to avoid large dt being multiplied by momentary gyro
readings causing a wild jump in orientation.  This problem may be
reduced when we start using the FIFO in the MPU6050 and detecting
FIFO overflows, but the overflows are best avoided in any case.

So take the chance to read the new orientations and properly wait for
the user to stop the movement in the previous axis before we start
integrating movement in new axis (or retry previous.)