main: Fine-tune AHRS parameters
Fine-tune accelerometer weight, IMU sample rate.  Lower divider produces
more samples and if we read all of them and average it seems we get
slightly better noise level than setting a lower sample rate -- it's not
clear from the spec if the MPU6050 IMU does the resampling when we lower
the sample rate -- it does seem to improve the noise level too, but not
by as much as receiving all of the samples and averaging them ourselves?
But this could be an artifact of the delay in reading the higher number
of samples from MPU6050, together with how we're measuring the standard
deviation in ahrs.c during calibrate.