ahrs: Reset q if too long since last update
Without this, if for whatever reason there has been a pause in the main
loop, the update will almost always cause a huge jump in orientation due
to us only having the latest gyro reading available.  That single
reading's noise and quantization error levels will be multiplied by the
huge dt and we effectively jump in a random direction.