motor: Split into separate classes for motor and driver
Create a separate, simpler, class for the BLDC motor driver with two
implementations: local PWM outputs (using SimpleFOC) and the
SBGC32_I2C_Drv for i2c-controlled motor driver.  On top of that we have
a motor class that does the actual control with velocity input and
voltage output to the driver object.  SimpleFOC now implements both
classes in one place but the full motor class is only there for
reference basically (maybe should #ifdef it out to save flash space).

Add comment to make it clear we now use SimpleFOC as a driver.

Some renaming.