mpu6050: Handle FIFO update race and improve overflow handling
First, account for the start of the FIFO being misaligned with the start
of the first sample in the FIFO after an overflow.  This is what was
screwing the AHRS up badly ever since I added the FIFO support for gyro
readings because the axes were effectively swapped.

Second, handle the (really easy to trigger) race condition between the
read and the MPU60x0 updating its FIFO registers.  This is easy but gets
complicated when combined with the FIFO overflow situation so the
handling of these two scenarios together is the difficult part here.

With these two improvements it's finally possible to compare AHRS
performance with and without the FIFO, and unfortunately the difference
isn't noticeable.  The performance is still poor where any movement of
around 90 degrees introduces a 6-8 degree error (if only fast enough
that the accelerometer Kp doesn't allow the accelerometer corrections to
catch up).  The hope is that it's going to be easier to find the source
of this error after an UI is added to visualise the real vs. estimated
orientations, which in any case would happen after a PoC level is
reached in actual camera stabilization.  And if that doesn't help then
more data collection and simulation will have to happen as the AHRS
precision is quite high priority.  It's at least 10x worse right now
than it should be.

Also make sure when there are no new samples we return the last sample
instead of 0's (in non-blocking mode).