axes: Make from_q optional in axes_q_to_step
We may already have (to_q x from_q^-1) in control.c where this is going
to be used so take that if possible.  In fact we might have the rotvec
but in global or main_imu frame of reference and would need to rotate it
still so perhaps easier to just take the q.