axes, main: Fix quaternion multiplication order
In two places we want the relative rotation from frame imu to main imu
but in the frame of reference of the frame imu, not global.  In other
words we want main imu orientation q expressed in the frame of reference
of the frame imu.

Switch around the multiplication operands.  Both places weren't active
code paths because of frame_ahrs not being in use.