control: Introduce basic control logic
Introduce a minimally functional skeleton for actual camera
stabilization.  This goes some way towards a proof of concept level but
at this point it rather disproves the concept ;-) as it works quite
poorly, mainly due to the BLDC motor feedback loop in motor-bldc.c (see
more below)

For now this code seems to work, it gathers the information about our
current orientation (the main and frame IMUs, or the camera's and the
handle's), applies various checks and calc to get the desired
orientation, calculates how far we are from where we want to be and how
fast we can get there (applying user's angular velocity and acceleration
limits in global frame), uses the joint axis calibration info to
calculate commands to each motor and eventually sends the new setpoints
to the motor driver PIDs.  It supports a boolean follow/no-follow switch
for each euler angle: typically users would enable yaw following and
disable pitch and roll following for the default gimbal use case.  It
also has minimal, untested switch to enable preserving yaw through the
moment of inverting the handle upside down (or over the gimbal-lock
position) so as to now flip the camera suddenly when this happens.

The main problem with this right now is the velocity-based motor control
in motor-bldc.c, which only has the encoder input as reference.  The
encoder resolution is way too low for maintaining smooth, slow rotations
speeds.  I have a planned improvement to map the IMU-provded rotation
(which will need both IMUs and some maths to get the relative speeds)
to each joint's axis and feed that to the BLDC drivers, but the upstream
firmware apparently doesn't do that or at least it doesn't need that
(because frame IMU is disabled by default) and still gets very good
per-joint speed control.  Likely the real solution here is to move to
torque-based control instead of velocity-based, which will need changes
both here and in motor-bldc.c and motor.h API.