control: Simplify conversion of final rotation to local frame
Not only to save cycles but also because at the 30 deg/s/s max
acceleration I decided to use as default (and others need like it
slower still) and starting at 0 speed, the conversion of the first
few steps from rotvec to quaternion somehow results in values close
to the closest a 32-bit float can come to 1.0, specifically the w
component which is the cosine of half the rotation angle.
Precision suffers and we risk the target speeds being rounded to 0
and no movement happening in some axes.  So reducing the number of
FP operations in those final steps should help.