main: Move all the non-critical code in loop() to idle callbacks
This is only code motion, no real changes are intended.  The idea is to
clean up loop() to make the real-time and critical elements of the loop
easy to see.  Also force some order on the individual bits as they're
now in individual functions and don't share variable scopes for example
which is good.