motor-bldc: Add motor_bldc_override_cur_omega()
Allow control.c to override the current angular velocity value that is
the input to the motor_bldc_loop() feedback loop.  This is in hopes that
control.c has a better estimate of it than what we can get by comparing
subsequent encoder readings and timestamps.  Our AS5600 encoders have
12-bit readings to cover the whole 360deg range and most other encoders
have that level of resolution.  We need an order of magnitude more for
smooth camera movements and more for other uses.