storage, main: Add FLASH storage API, commands for reading&writing
Add storage.cpp/.h (cpp because we'll need TwoWire which is C++ only)
which currently supports the STM32 on-chip flash as the backend, allows
saving and loading data in a versioned structure.

In main, initialize the storage backend and load config from storage on
start, add read&write commands ('r' and 'w') for loading and saving
on-demand.

The downside of using the internal flash is that 1. it's rated for fewer
program cycles than the EEPROM and 2. all of it is erased by stm32ld
when uploading new code.