motor-bldc: Fix reading encoder values during calibration
The main loop isn't running so encoder_update() isn't being called.  We
need to call it here, or we should have maybe left this function alone
when we converted it to the encoder updates happening in the main loop.

While there also start using the new error_print() to output calibration
error messages.