main: Handle CMD_EXECUTE_MENU as sent by PilotFly RM-1 remote
Handle a bunch of the command/key IDs that can be signalled by this
SimpleBGC serial API message, even though it's a small percentage of all
the defined keys.

It seems that the setting the camera position with
control.target_ypr_offsets is correct despite the upcoming change of how
we communicate the speed/angle commands to control.c.  It should work in
speed mode but it's unclear what the expected result is with either RC
or serial API signals in angle mode.

We could maybe instead be writing to control.sbgc_api_ypr_offsets and set
the override start timestamps, this would override the RC in signals
idependent of the RC input speed/angle mode... for a short time.