axes: Precalculate the axes Jacobian in each iteration
Update the Jacobian in axes_precalc_rel_q(), before control_step() is
called.  control_step() will be using it more than once so it makes
sense to move it out of axes_rotvec_to_step_proj() and also move out the
velocity calculation we added in there just to reuse the Jacobian value
that this function already had.