main: Rename i2c_aux to i2c_int to match orig firmware naming
The serial API docs call the two busses "external" and "internal" (the
firmware code possibly calls them line A and line B judging from the
ext_i2c_drv board docs), we called them "main" and "aux" for lack of
better idea.  Rename "aux" to "int" to make it more clear to the users
of the serial api which bus is which.  Probably enough for just one of
the two names to ~match.