motor-pwm: Fix voltage scale when handling .set_phase_voltage
There was a mismatch in motor strength between the local (onboard)
driver and the SBGC_I2C_Drv drive boards due to the API difference.
This was because SimpleFOC was dividing the received values by what it
thought is current voltage, multiply the inputs by the same value before
passing to SimpleFOC to remove this effect.

Adjust default PID params in main.cpp to match.