axes: Tentative code for mechanical limit avoidance
This function only stops control_step from crossing into the forbidden
zone around each axis' mechanical end-stops or whatever was configured
as min and max limits by user (and we have no UI for that right now).

Eventually we want to be able to decide to go full circle the other
direction to arrive at the target orientation avoiding the forbidden
range of angles.