control: Factor out the orientation interpolation-based path controller
Intended as a pure refactoring commit that shouldn't have functional
side effects, in an attempt to make the addition of a new path
controller (in the next commit) more reviewable.

Note that control_calc_path_step_orientation() multiplies the final
angles by dt and the caller (control_step) then undoes this.  A little
annoying but it justifies the "step" in the function name and should
be causing any harm.