motor-pwm,moremath: Normalize angle before .setPhaseVoltage()
The _sincos implementation in Simple FOC needs angles between 0-2*Pi.
Move angle_normalize_0_2pi from axes.c to moremath.h and change it so
that it should in theory never reduce the float resolution by adding
and then subtracting from it.  This comes at the cost of an extra branch
though.

Apply the same logic in angle_normalize_pi.