axes,main: Choose the inner angle closest to home value, not to 0
When calibrating home position also save the joint angles.  When
choosing between one of the two solutions in axes_q_to_angles_*(),
select the preferred solution based on this info instead of hardcoding
the -90 - +90deg range as the angles are not referenced to any specific
geometrical feature and depend on the encoder mounting at factory.

We probably also want to sometimes pass the current angles instead of
home angles depending either on a user setting or distance between
current and target (TODO)