params,main: Introduce runtime parameters, add list
Start adding a slightly more flexible mechanism for writing and reading
the runtime parameters and/or state for our code.  Each bit of
state/config/etc. is assigned a unique and permanent ID.  The IDs never
change for future compatibility, but they may be deprecated/removed and
reserved, never reused.

The idea is that this be used both for the companion applications API --
in fact writing and reading the params could provide almost all of the
API that's ever needed -- and a new storage format.  This would eventually
replace the current non-updatable big hardcoded struct format.

This whole infrastructure is just a starting point for now and not used.
It may however already start slightly shaping some of the future code
changes by adding restrictions on moving and renaming things (by
triggering compile errors) thus adding some order.

Until the IDs are exposed in the API or storage and new rules for them
start being enforced, I plan to change them in incompatible ways to keep
the parameters list ordered while I can.

At the current number of parameters this adds about 1.3kB to flash size
(except the linker currently realizes it's not used and drops the
parameter table from the final binary) and 0 bytes of RAM use as
everything is static and doesn't need to be copied from flash to RAM.

Platformio's build system doesn't seem to be tracking changes in files
ending in .inc so it will not automatically rebuild main.o when
parameters code changes, until main.cpp timestamp changes too.