i2c: Work around FlexWire forcing its own Wire.h on us
With FlexWire installed we'd be including its own version of Wire.h
which also defined the TwoWire identifier, instead of the Arduino (in
our case stm32 framework's) shipped TwoWire class.  The result was that
both of our i2c busses were using FlexWire while the idea was to use the
faster and slightly more capable hardware implementation for the
"external" bus -- the one which the sensors, the encoder and the motors
are on -- and FlexWire bitbanging for the "internal" bus, i.e. the one
that we only use occasionally, not in every loop iteration.

Opened https://github.com/felias-fogg/FlexWire/issues/10 to signal this
can be a problem.