mpu6050: Fix USE_FIFO
When I moved things around, the #define USE_FIFO was actually after the
ifdef USE_FIFO in the read function so the FIFO wasn't being read, even
though it was being enabled in the initializer.

Also use a macro for the max transfer size which depends on which
library we're using.  If we'd really been using the FIFO, it wouldn't
have worked with FlexWire because it can handle only up to 32 bytes per
transfer.  Now that we switched back to TwoWire we can keep using 255.