main, ahrs: Save timestamp at mainloop stages for perf tuning
Add the '-' key command for reading the timers.  They tell you basically
how many microseconds have passed at the different stages of the main
loop iteration from the beginning of that iteration.  From there you
have to figure out on your own which step did what.  In any case, this
mainly reveals that I2C communication takes 80% of our time.

A minor issue is that the final two counter values are from the previous
loop because the '-' key is being handled within the main loop, towards
the end in the idle callbacks.  The final two timestamps haven't been
written yet as we're sending the values to the serial port.