main, mpu6050: Stop using the FIFO, enable DLPF
Reading more than the last sample from the FIFO turned out to have
little benefit to tracking accuracy -- with and without it, the current
code, on my 2 device units both give me errors up to 3-4 degrees after a
90 degree turn (with accelerometer as reference) so the averaging, while
it greatly reduced the noise in the readings, wasn't all that useful.
It was also very slow because we'd be reading about 4 samples per loop
instead of 1, but apart from that we also needed multiple trasnfers to
first read the FIFO counter and then read the gyro data from the FIFO
separately from the accelerometer data so this was all compounding to
about 1.8ms per loop with hardware I2C at 400kHz (we can't crank it up
anymore).  It was dependent on the main loop rate though.  With the main
loop roughly tuned with the MPU6050 sample rate it would take about
0.8-0.9 ms (depending on whether there has been 1 or 2 samples in the
FIFO in a given iteration) but obivously without any of the benefits of
the averaging because in the end we were only reading a single sample.

Without the FIFO the read takes 0.4ms consistently.

Enable the DLPF to get *some* amount of noise reduction on the chip
side, although again it doesn't seem to improve the tracking accuracy
noticeable.  Current thinking is the accuracy problem stems from sensor
axis misalignment so more work needed.