control: Split out the joint velocity calc for motor PIDs
Split the code estimating current rotation velocity per joint and
providing this as feedback to the BLDC driving PID loops, out from
control_step() whose purpose is hardly related to this.  Maybe should
even move this out of control.c but it's convinient there for now
because we use some the same inputs and the velocity output in the
control loop.