main: Crank up main loop rate by 3x+
Boost the target rate from 128Hz to 400Hz (512 would work too although
might be tight, but realized there's no reason to have this at a power
of 2).

This is mainly to let the motor PID loops and the FOC driving respond
to feedback faster and be able to send shorter pulses to the motors.
Initial idea was to have a slow main loop with maybe interrupt-driven
"RT" or "high-priority" tasks at a higher rate, i.e. fractional loop
iterations.  It was a bigger mess to do it that way though, speically
with locking for I2C and other globals, it would require queues and what
not.

Instead run the loop at the higher rate and run the "low priority"
tasks, i.e. control_step() only every 4 iterations, together with its
inputs such as the encoder reading and axes orientations calc.  (These
are actually used by control_update_joint_vel() which is part of the
motor feedback loop... but the accuracy is not as critical for this task
so we're Ok with the slighlty outdated encoder readings.)

Spread these different "low-priority" tasks between the different loop
iterations so as not to have 3 fast and then one long iteration.
Instead iterations 1, 2 and 3 all perform parts of the slower tasks.

This may be a bit overkill, so we may define the EVERY_4_CYCLES() macro
to 1 for simplicity if it becomes any problem.  We can also freely go
back to a slower loop rate if needed because in testing the faster rate
doesn't seem to provide the expected benefit for the motor driving so far.
I.e. the motors enter oscillations at about the same P, I or D values as
before so we cannot seem to set them noticeably higher than before,
which was the goal.