main: Add mechanism to bail to bootloader early
In order to avoid having to use the FLASH jumper during development we
had a way to go to the bootloader from the main loop by user request and
from the crash handler if things went wrong.  But if any of the setup or
main loop steps had caused a lockup/inifinite loop, we'd be out of luck.
For non-developers that was covered by only installing tested releases I
guess, where init was expected to complete fine (although perhaps a bad
config could get them stuck too).

With this you can press 'q' in the early boot to abort the firmware and
go directly to the bootloader to reflash and fix any bug causing a boot
time lockup, such as the MPU6050 writing data into its FIFO at a higher
rate than we can possibly read over I2C getting us stuck reading I2C
ifinitely.  Now, bugs that require the use of the FLASH jumper, to force
bootloader, are going to be unlikely.