ahrs: Accept sensor axis alignment+sensitivity matrices
Multiply readings by these matrices for both accelerometer and gyro for
future use when we can actually calculate any level of calibration info
for axis alignment.  These sensors *seem* to have an axis misalignment
that is possibly the main source of our poor tracking ability, but hard
to confirm yet.  There are a few papers that provide not-too-difficult
calibration algorithms but in any case they're too complicated to do in
the firmware so we're just leaving space here for the API to let
external scripts and/or app do the hard work and give us the final data.
I have no working code for any of that at this time.