control: Fix possible division by 0, harden the recovery
delta_angle was occasionally 0 and we'd end up with infinity in one of
the velocity_vec components (which would be replaced by a nan in the
following steps) and the whole loop would stop working.  Also the
recovery mechanism that was there wasn't actually fixing this case (was
only printing warnings) so improve that.