motor-bldc: Add param to trust current Kp value
Add a way to use the requested target velocity from previous step as new
base velocity for error calculation.  This basically says we're trusting
the PID parameters are doing what they should be doing and we can ignore
part of the feedback from sensors and replace that part by this guessed
velocity.  The effect of this is intended to be a reduction in the
amount of sensor noise and feedback delay getting through to the output
and causing delay and noise in the motor position.

This sounds questionable but it does seem to allow use to crank up the P
by some amount and lower the D factors.  And the improvement seems to be
real, not simply a way to wash away some of the value of the parameters
so they need to be cranked up, and neither it should be at the cost of a
delay which is often the case when something seems to be improving the
PID tracking: it either makes the PID slower or the params are
artificially boosted but without a real gain.

More work is needed nevertheless as the velocity tracking is still not
perfect.

Drop debug code that was there by mistake since ac4071042473b710e82f7a34e17206009d9aca71