control: First attempt at stability at gimbal lock
See long code comment.  Doesn't work incredibly well but there is an
improvement.  This is a workaround really, work needed on the proper
solution.  But the solution would be only for the SHORT path type so we
may want to keep this for the other path types, and probably need to do
the same thing in control_update_joint_vel()?