axes: Calculate a 2D Jacobian near gimbal lock
Check if we're close to gimbal lock, emit debug warnings and change the
way the jacobian is calculated in a way that we should in theory never
see a 0 determinant.  Use a direct inverse instead of the more
complicated pseudo-inverse when we know we're not close to the gimbal
lock.  If we are close to gimbal lock, use basically only two of the
axis vectors as Jacobian columns, invert the resulting 2x3 matrix.