control: Gradually attenuate inputs from ~30deg from gimbal lock
Try to implement a gradual transition from the normal orientation
processing with the full Jacobian inverse to the gimbal lock special
case to get a smoother movement.  I still occiasionally see the motors
go out of control briefly (big oscillations) when *exiting* the gimbal
lock zone, may be an artifact of poor PIDs (poor calibration and poor
feedback) and jumps in inputs.  Everything is good in slow movement.