ahrs: Add IMU readings to state and expose as API params
Expose latest IMU readings for the benefit of potential calibration
scripts.  Ideally we'd let the script request the average or sum of the
readings between two points in time but for now let's see if we get away
with just this.  Any script will have to read these as often as possible
and extrapolate / average on their own, probably needing slower motion
and more static time between movements in the accelerometer averaging
phases.